ref: move AttachmentCarousel to TS
