+++ src/components/Attachments/AttachmentCarousel/index.tsx
@@ -1,32 +1,42 @@
 import lodashGet from 'lodash/get';
 import React, {useCallback, useEffect, useRef, useState} from 'react';
-import {Keyboard, PixelRatio, View} from 'react-native';
+import {FlatList, Keyboard, PixelRatio, View} from 'react-native';
 import {withOnyx} from 'react-native-onyx';
 import _ from 'underscore';
 import BlockingView from '@components/BlockingViews/BlockingView';
 import * as Illustrations from '@components/Icon/Illustrations';
 import withLocalize from '@components/withLocalize';
+import withWindowDimensions from '@components/withWindowDimensions';
 import compose from '@libs/compose';
+import * as DeviceCapabilities from '@libs/DeviceCapabilities';
 import Navigation from '@libs/Navigation/Navigation';
 import useThemeStyles from '@styles/useThemeStyles';
 import variables from '@styles/variables';
 import ONYXKEYS from '@src/ONYXKEYS';
+import AttachmentCarouselCellRenderer from './AttachmentCarouselCellRenderer';
 import {defaultProps, propTypes} from './attachmentCarouselPropTypes';
+import CarouselActions from './CarouselActions';
 import CarouselButtons from './CarouselButtons';
-import CarouselItem from './CarouselItem.tsx';
+import CarouselItem from './CarouselItem';
 import extractAttachmentsFromReport from './extractAttachmentsFromReport';
-import AttachmentCarouselPager from './Pager';
 import useCarouselArrows from './useCarouselArrows';
 
-function AttachmentCarousel({report, reportActions, parentReportActions, source, onNavigate, setDownloadButtonVisibility, translate, transaction, onClose}) {
-    const styles = useThemeStyles();
-    const pagerRef = useRef(null);
+const viewabilityConfig = {
+    // To facilitate paging through the attachments, we want to consider an item "viewable" when it is
+    // more than 95% visible. When that happens we update the page index in the state.
+    itemVisiblePercentThreshold: 95,
+};
 
-    const [containerDimensions, setContainerDimensions] = useState({width: 0, height: 0});
+function AttachmentCarousel({report, reportActions, parentReportActions, source, onNavigate, setDownloadButtonVisibility, translate, transaction}) {
+    const styles = useThemeStyles();
+    const scrollRef = useRef(null);
+
+    const canUseTouchScreen = DeviceCapabilities.canUseTouchScreen();
+
+    const [containerWidth, setContainerWidth] = useState(0);
     const [page, setPage] = useState(0);
     const [attachments, setAttachments] = useState([]);
     const [activeSource, setActiveSource] = useState(source);
-    const [isPinchGestureRunning, setIsPinchGestureRunning] = useState(true);
     const [shouldShowArrows, setShouldShowArrows, autoHideArrows, cancelAutoHideArrows] = useCarouselArrows();
     const [isReceipt, setIsReceipt] = useState(false);
 
@@ -62,7 +72,7 @@ function AttachmentCarousel({report, reportActions, parentReportActions, source,
             }
         }
         // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [reportActions, compareImage]);
+    }, [reportActions, parentReportActions, compareImage]);
 
     /**
      * Updates the page state when the user navigates between attachments
@@ -70,19 +80,25 @@ function AttachmentCarousel({report, reportActions, parentReportActions, source,
      * @param {number} index
      */
     const updatePage = useCallback(
-        (newPageIndex) => {
+        ({viewableItems}) => {
             Keyboard.dismiss();
-            setShouldShowArrows(true);
 
-            const item = attachments[newPageIndex];
+            // Since we can have only one item in view at a time, we can use the first item in the array
+            // to get the index of the current page
+            const entry = _.first(viewableItems);
+            if (!entry) {
+                setIsReceipt(false);
+                setActiveSource(null);
+                return;
+            }
 
-            setPage(newPageIndex);
-            setIsReceipt(item.isReceipt);
-            setActiveSource(item.source);
+            setIsReceipt(entry.item.isReceipt);
+            setPage(entry.index);
+            setActiveSource(entry.item.source);
 
-            onNavigate(item);
+            onNavigate(entry.item);
         },
-        [setShouldShowArrows, attachments, onNavigate],
+        [onNavigate],
     );
 
     /**
@@ -91,39 +107,61 @@ function AttachmentCarousel({report, reportActions, parentReportActions, source,
      */
     const cycleThroughAttachments = useCallback(
         (deltaSlide) => {
-            const nextPageIndex = page + deltaSlide;
-            updatePage(nextPageIndex);
-            pagerRef.current.setPage(nextPageIndex);
+            const nextIndex = page + deltaSlide;
+            const nextItem = attachments[nextIndex];
 
-            autoHideArrows();
+            if (!nextItem || !scrollRef.current) {
+                return;
+            }
+
+            scrollRef.current.scrollToIndex({index: nextIndex, animated: canUseTouchScreen});
         },
-        [autoHideArrows, page, updatePage],
+        [attachments, canUseTouchScreen, page],
+    );
+
+    /**
+     * Calculate items layout information to optimize scrolling performance
+     * @param {*} data
+     * @param {Number} index
+     * @returns {{offset: Number, length: Number, index: Number}}
+     */
+    const getItemLayout = useCallback(
+        (_data, index) => ({
+            length: containerWidth,
+            offset: containerWidth * index,
+            index,
+        }),
+        [containerWidth],
     );
 
     /**
      * Defines how a single attachment should be rendered
-     * @param {{ reportActionID: String, isAuthTokenRequired: Boolean, source: String, file: { name: String }, hasBeenFlagged: Boolean }} item
+     * @param {Object} item
+     * @param {String} item.reportActionID
+     * @param {Boolean} item.isAuthTokenRequired
+     * @param {String} item.source
+     * @param {Object} item.file
+     * @param {String} item.file.name
+     * @param {Boolean} item.hasBeenFlagged
      * @returns {JSX.Element}
      */
     const renderItem = useCallback(
-        ({item, isActive}) => (
+        ({item}) => (
             <CarouselItem
                 item={item}
-                isFocused={isActive && activeSource === item.source}
-                onPress={() => setShouldShowArrows(!shouldShowArrows)}
+                isFocused={activeSource === item.source}
+                onPress={canUseTouchScreen ? () => setShouldShowArrows(!shouldShowArrows) : undefined}
             />
         ),
-        [activeSource, setShouldShowArrows, shouldShowArrows],
+        [activeSource, canUseTouchScreen, setShouldShowArrows, shouldShowArrows],
     );
 
     return (
         <View
             style={[styles.flex1, styles.attachmentCarouselContainer]}
-            onLayout={({nativeEvent}) =>
-                setContainerDimensions({width: PixelRatio.roundToNearestPixel(nativeEvent.layout.width), height: PixelRatio.roundToNearestPixel(nativeEvent.layout.height)})
-            }
-            onMouseEnter={() => setShouldShowArrows(true)}
-            onMouseLeave={() => setShouldShowArrows(false)}
+            onLayout={({nativeEvent}) => setContainerWidth(PixelRatio.roundToNearestPixel(nativeEvent.layout.width))}
+            onMouseEnter={() => !canUseTouchScreen && setShouldShowArrows(true)}
+            onMouseLeave={() => !canUseTouchScreen && setShouldShowArrows(false)}
         >
             {page === -1 ? (
                 <BlockingView
@@ -135,7 +173,7 @@ function AttachmentCarousel({report, reportActions, parentReportActions, source,
             ) : (
                 <>
                     <CarouselButtons
-                        shouldShowArrows={shouldShowArrows && !isPinchGestureRunning}
+                        shouldShowArrows={shouldShowArrows}
                         page={page}
                         attachments={attachments}
                         onBack={() => cycleThroughAttachments(-1)}
@@ -144,29 +182,45 @@ function AttachmentCarousel({report, reportActions, parentReportActions, source,
                         cancelAutoHideArrow={cancelAutoHideArrows}
                     />
 
-                    {containerDimensions.width > 0 && containerDimensions.height > 0 && (
-                        <AttachmentCarouselPager
-                            items={attachments}
+                    {containerWidth > 0 && (
+                        <FlatList
+                            keyboardShouldPersistTaps="handled"
+                            listKey="AttachmentCarousel"
+                            horizontal
+                            decelerationRate="fast"
+                            showsHorizontalScrollIndicator={false}
+                            bounces={false}
+                            // Scroll only one image at a time no matter how fast the user swipes
+                            disableIntervalMomentum
+                            pagingEnabled
+                            snapToAlignment="start"
+                            snapToInterval={containerWidth}
+                            // Enable scrolling by swiping on mobile (touch) devices only
+                            // disable scroll for desktop/browsers because they add their scrollbars
+                            // Enable scrolling FlatList only when PDF is not in a zoomed state
+                            scrollEnabled={canUseTouchScreen}
+                            ref={scrollRef}
+                            initialScrollIndex={page}
+                            initialNumToRender={3}
+                            windowSize={5}
+                            maxToRenderPerBatch={3}
+                            data={attachments}
+                            CellRendererComponent={AttachmentCarouselCellRenderer}
                             renderItem={renderItem}
-                            initialIndex={page}
-                            onPageSelected={({nativeEvent: {position: newPage}}) => updatePage(newPage)}
-                            onPinchGestureChange={(newIsPinchGestureRunning) => {
-                                setIsPinchGestureRunning(newIsPinchGestureRunning);
-                                if (!newIsPinchGestureRunning && !shouldShowArrows) {
-                                    setShouldShowArrows(true);
-                                }
-                            }}
-                            onSwipeDown={onClose}
-                            containerWidth={containerDimensions.width}
-                            containerHeight={containerDimensions.height}
-                            ref={pagerRef}
+                            getItemLayout={getItemLayout}
+                            keyExtractor={(item) => item.source}
+                            viewabilityConfig={viewabilityConfig}
+                            onViewableItemsChanged={updatePage}
                         />
                     )}
+
+                    <CarouselActions onCycleThroughAttachments={cycleThroughAttachments} />
                 </>
             )}
         </View>
     );
 }
+
 AttachmentCarousel.propTypes = propTypes;
 AttachmentCarousel.defaultProps = defaultProps;
 AttachmentCarousel.displayName = 'AttachmentCarousel';
@@ -196,4 +250,5 @@ export default compose(
         },
     }),
     withLocalize,
+    withWindowDimensions,
 )(AttachmentCarousel);
